# replace the mutex locks in logging with atomic bool for the "on" flag